Make fetcher propagate 404s
So EventFinder can detect missing streams.  Before, it would retry them.

This may cause a problem for the availability service which makes use of
streams with date-based templated stream names: Bob suggests availability
relies on this to retry until a stream gets created.  We decided to leave this
for when availability needs to update its AP version.